Exit status

+++ srcs/builtins/cd.c
@@ -36,7 +36,7 @@ static void	update_oldpwd(t_data *data)
  * If PWD is not unset, the value will be changed as well.
  * data->pwd exists in case PWD doesn't, since pwd should still work.
 */
-int8_t	cd(t_data *data, t_group *group)
+uint32_t	cd(t_data *data, t_group *group)
 {
 	t_env	*env;
